Move polars tests to a separate file
